LayoutTests/http/tests/loading/preload-image-srcset.html

<html>
<head>
<script src="http://127.0.0.1:8000/resources/slow-script.pl?delay=1000"></script>
<script src="../resources/js-test-pre.js"></script>
</head>
<body>
<script>
    shouldBeFalse("internals.isPreloaded('resources/image1.png');");
    shouldBeTrue("internals.isPreloaded('resources/base-image1.png');");
    shouldBeTrue("internals.isPreloaded('resources/image2.png');");
</script>
<div>This test passes if the fallback image (defined in the src attribute) is not preloaded, but the 1x version (defined in srcset) is.</div>
<div>The second image is here to make sure isPreloaded is working properly</div>
<img height="100" width="100" src="resources/preload-test.jpg" srcset="resources/base-image1.png 1x, resources/base-image3.png 3x, resources/base-image2.png 2x"></img>
<img src="resources/image2.png">
</body>
</html>
